Start talking to the CC1101.
# Testing

I am repeatedly sending the SNOP (0x3D) command strobe
to get the status byte. The result is as expected.

## For a "write" access

    while sleep .05; do printf "\x3D" >/dev/ttyUSB0; done
    
We we get back 0x0F (IDLE state, 15+ bytes available in TX FIFO).
This is as expected.

## For a "read" access

    while sleep .05; do printf "\xBD" >/dev/ttyUSB0; done
    
We we get back 0x00 (IDLE state, 0 bytes available in RX FIFO).
This is as expected.