"first" attempt at testing
Probably not a meaningful test, but the main idea here is that I
previously only had this "test" in the commit message and in my notes,
whereas now I have an automated way to run it.

In the future when I have something that I actually want to test, this
will be a good quickstart.