Get the radio working finally.
I don't have a spectrum analyzer available. All I have is a Rigol 1054Z
(100MHz analog bandwidth).
But remember that the nominal bandwidth of a scope is generally the -3dB
bandwidth. The 433MHz carrier can still get through (albeit highly
attenuated) and is visible on the smaller voltage ranges.

One needs to consider the scope's sample rate. For example, the 1054Z is
1 GSa/s and the way I have the scope configured to trigger and all right
now ends up only getting 500MSa/s on the waveform for that channel
(which it conveniently shows on the screen; great feature!). This means
that the 433MHz carrier shows up at 500MHz - 433MHz = 67MHz.
Disabling the other channel that was using up the sample rate so that I
get the full 1 GSa/s I can see the unaliased 433MHz carrier.

Test procedure:
1. Run the MSP432
2. Run test.py (just using this to repeatedly trigger a transmission;
sort of hacky; piggy-backing on some stuff I had there before)
3. Probe around the CC1101 to see an "incrementing" pattern of bursts.

Note that the "0" is a tone burst ("on"), while "1" is a dead time
("off"). So the code uses `~iters` so that on the scope it will appear
in the more intuitive "presence of a thing means 1; absence means 0" so
that the incrementing pattern is obvious.
Looking on the scope, the bytes are sent MSB first.

From my notes, the CC1101 is configured as follows:
- 433 MHz
- TX power is 0dBm
- 10kBaud for a symbol period of 100us.
- ASK/OOK