Switch up the pre/post conditions of uart_send()
It makes more sense for uart_send to block until the byte is sent. The
original behavior I had here was I think notionally trying to keep the
TXIFG always equal to 0, which isn't really relevant for a blocking
interface (it is what you want to strive for in an interrupt-driven
interface though).

Also, clarify the comments on uart_recv/uart_send.