Initial playing around with frequencies.
Right now, I'm mostly interested in which frequencies the PLL can lock
at. Some previous sloppy experiments made it seem like the range was
pretty limited, but I want to take a closer and more careful look.

Main thing is adding an 'f' command (vaguely stands for "frequency").
I expect to change the meaning of this command as I get a clearer idea
about what is useful.
Right now it just sets FREQ2 to the given value, and returns the value
of FSCAL1 (according to the CC1101 datasheet: "The PLL is in lock if the
register content is different from 0x3F"; need to test this statement
myself though).